+++ src/CoreShop/Component/Core/Cart/Rule/Condition/CountriesConditionChecker.php
@@ -8,41 +8,36 @@
  *
  * @copyright  Copyright (c) 2015-2017 Dominik Pfaffenbauer (https://www.pfaffenbauer.at)
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
-*/
+ */
 
 namespace CoreShop\Component\Core\Cart\Rule\Condition;
 
 use CoreShop\Component\Address\Model\AddressInterface;
 use CoreShop\Component\Core\Model\CountryInterface;
 use CoreShop\Component\Customer\Model\CustomerInterface;
+use CoreShop\Component\Order\Cart\Rule\Condition\AbstractConditionChecker;
 use CoreShop\Component\Order\Model\CartInterface;
-use CoreShop\Component\Rule\Condition\ConditionCheckerInterface;
-use Webmozart\Assert\Assert;
+use CoreShop\Component\Order\Model\CartPriceRuleVoucherCodeInterface;
 
-final class CountriesConditionChecker implements ConditionCheckerInterface
+final class CountriesConditionChecker extends AbstractConditionChecker
 {
     /**
      * {@inheritdoc}
      */
-    public function isValid($subject, array $configuration)
+    public function isCartRuleValid(CartInterface $cart, CartPriceRuleVoucherCodeInterface $voucher, array $configuration)
     {
-        /**
-         * @var $subject CartInterface
-         */
-        Assert::isInstanceOf($subject, CartInterface::class);
-
-        if (!$subject->getCustomer() instanceof CustomerInterface) {
+        if (!$cart->getCustomer() instanceof CustomerInterface) {
             return false;
         }
 
-        if (!$subject->getInvoiceAddress() instanceof AddressInterface) {
+        if (!$cart->getInvoiceAddress() instanceof AddressInterface) {
             return false;
         }
 
-        if (!$subject->getInvoiceAddress()->getCountry() instanceof CountryInterface) {
+        if (!$cart->getInvoiceAddress()->getCountry() instanceof CountryInterface) {
             return false;
         }
 
-        return in_array($subject->getInvoiceAddress()->getCountry()->getId(), $configuration['countries']);
+        return in_array($cart->getInvoiceAddress()->getCountry()->getId(), $configuration['countries']);
     }
 }