+++ src/CoreShop/Component/Order/Cart/Rule/CartPriceRuleOrderProcessorInterface.php
@@ -14,17 +14,18 @@ namespace CoreShop\Component\Order\Cart\Rule;
 
 use CoreShop\Component\Order\Model\CartInterface;
 use CoreShop\Component\Order\Model\CartPriceRuleInterface;
+use CoreShop\Component\Order\Model\CartPriceRuleVoucherCodeInterface;
 use CoreShop\Component\Order\Model\SaleInterface;
 
 interface CartPriceRuleOrderProcessorInterface
 {
     /**
      * @param CartPriceRuleInterface $cartPriceRule
-     * @param $usedCode
+     * @param CartPriceRuleVoucherCodeInterface $voucherCode
      * @param CartInterface $cart
      * @param SaleInterface $sale
      *
      * @return mixed
      */
-    public function process(CartPriceRuleInterface $cartPriceRule, $usedCode, CartInterface $cart, SaleInterface $sale);
+    public function process(CartPriceRuleInterface $cartPriceRule, CartPriceRuleVoucherCodeInterface $voucherCode, CartInterface $cart, SaleInterface $sale);
 }