+++ src/CoreShop/Component/Order/Cart/Rule/CartPriceRuleUnProcessor.php
@@ -14,6 +14,7 @@ namespace CoreShop\Component\Order\Cart\Rule;
 
 use CoreShop\Component\Order\Model\CartInterface;
 use CoreShop\Component\Order\Model\CartPriceRuleInterface;
+use CoreShop\Component\Order\Model\CartPriceRuleVoucherCodeInterface;
 use CoreShop\Component\Order\Model\ProposalCartPriceRuleItemInterface;
 
 use CoreShop\Component\Registry\ServiceRegistryInterface;
@@ -38,7 +39,7 @@ class CartPriceRuleUnProcessor implements CartPriceRuleUnProcessorInterface
     /**
      * {@inheritdoc}
      */
-    public function unProcess(CartPriceRuleInterface $cartPriceRule, $usedCode, CartInterface $cart)
+    public function unProcess(CartPriceRuleInterface $cartPriceRule, CartPriceRuleVoucherCodeInterface $voucherCode, CartInterface $cart)
     {
         $priceRuleItem = null;
 
@@ -48,7 +49,7 @@ class CartPriceRuleUnProcessor implements CartPriceRuleUnProcessorInterface
                     $cartsRule = $rule->getCartPriceRule();
 
                     if ($cartsRule instanceof CartPriceRuleInterface) {
-                        if ($cartsRule->getId() === $cartPriceRule->getId() && $usedCode === $rule->getVoucherCode()) {
+                        if ($cartsRule->getId() === $cartPriceRule->getId() && $voucherCode->getCode() === $rule->getVoucherCode()) {
                             $priceRuleItem = $rule;
                             break;
                         }